Disable display bus due to footprint error
The footprint of the 74LV273 octal flip-flop is wrong and doesn't
match that of the actual chip. As a result, the flip-flops cannot
be used and the bus must be set to high impedance for now.

Next step might be an adaptor board that can be soldered on top of
the footprint.